[GrB] Extract solution to class
