+++ solutions/GraphBLAS/utils.h
@@ -1,6 +1,8 @@
 #pragma once
 
 #include <stdexcept>
+#include <vector>
+#include <chrono>
 
 extern "C" {
 #include <GraphBLAS.h>
@@ -16,6 +18,21 @@ struct BenchmarkParameters {
     std::string Query;
     unsigned long thread_num = 1;
 };
+
+BenchmarkParameters parse_benchmark_params();
+
+using score_type = std::tuple<uint64_t, time_t, GrB_Index>;
+
+struct BenchmarkPhase {
+    static const std::string Initialization;
+    static const std::string Load;
+    static const std::string Initial;
+    static const std::string Update;
+};
+
+void report(const BenchmarkParameters &parameters, int iteration, const std::string &phase,
+            std::chrono::nanoseconds runtime,
+            std::optional<std::vector<score_type>> result_reversed_opt = std::nullopt);
 
 void WriteOutDebugMatrix(const char *title, GrB_Matrix result);
